requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-github>=4.0.0,<5.0.0
hocon